Do axis scaling in AnalogControl, enabling 3D pad
We actually have two types of analog axes to deal with: two-sided which
can be positive or negative (eg wheel, joystick) and one-sided
(triggers, pedals). This patch adds a flag to AnalogControl to determine
the appropriate scaling and also then applies this to the AnalogPad (3D
Control Pad) to get the triggers working.