Allow digital controls to be set by analog axes and vice versa
This makes the set(bool) and set(int16_t) methods a part of the Control
ABC. Digital controls apply a hysteresis on the axis value when fed an
int, while analog controls just go hard on and off when fed a bool.

The 3D Pad's analog triggers also set the digital trigger bits with some
hysteresis. This change lets us implement that by additionally binding
the axis to the digital trigger buttons.

This also gets rid of a bunch of ugly static_casts around the place.